# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CertEfficiency API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /study-plans:
    post:
      operationId: createStudyPlan
      tags: [studyPlans]
      summary: Generate a study plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyPlanInput"
      responses:
        "201":
          description: Study plan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyPlan"
  /study-plans/current:
    get:
      operationId: getCurrentStudyPlan
      tags: [studyPlans]
      summary: Get current active study plan
      responses:
        "200":
          description: Current study plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyPlan"
        "404":
          description: No active study plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /study-plans/{id}:
    patch:
      operationId: updateStudyPlan
      tags: [studyPlans]
      summary: Update study plan details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyPlanUpdate"
      responses:
        "200":
          description: Updated study plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyPlan"
...
  /uploads:
    get:
      operationId: listUploads
      tags: [uploads]
      summary: List user uploads
      responses:
        "200":
          description: List of uploads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Upload"
    post:
      operationId: createUpload
      tags: [uploads]
      summary: Upload a file (multipart/form-data handled by server middleware)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadMetadataInput"
      responses:
        "201":
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Upload"
        "400":
          description: Unsupported file type or size exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /uploads/{id}:
    delete:
      operationId: deleteUpload
      tags: [uploads]
      summary: Delete an upload
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # Auth schemas
...
    StudyPlanInput:
      type: object
      required: [certificationId]
      properties:
        certificationId:
          type: string
        examDate:
          type: ["string", "null"]
          format: date
        weeklyHoursAvailable:
          type: integer
    StudyPlan:
      type: object
      required: [id, userId, certificationId, status, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        certificationId:
          type: string
        certificationName:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, completed, superseded]
        examDate:
          type: ["string", "null"]
          format: date
        daysRemaining:
          type: ["integer", "null"]
        weeklyHoursAvailable:
          type: ["integer", "null"]
        weakDomains:
          type: array
          items:
            type: string
        strongDomains:
          type: array
          items:
            type: string
        milestones:
          type: array
          items:
            type: object
            additionalProperties: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/StudyPlanItem"
        createdAt:
          type: string
          format: date-time
    StudyPlanItem:
      type: object
      required: [id, studyPlanId, title, scheduledDate, durationMinutes, itemType, completed]
      properties:
        id:
          type: string
        studyPlanId:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        scheduledDate:
          type: string
          format: date
        durationMinutes:
          type: integer
        itemType:
          type: string
          enum: [study, review, mock_exam, milestone]
        domain:
          type: ["string", "null"]
        completed:
          type: boolean
        completedAt:
          type: ["string", "null"]
          format: date-time
    StudyPlanUpdate:
      type: object
      properties:
        weeklyHoursAvailable:
          type: integer
        examDate:
          type: string
          format: date

    # Progress schemas
...
    UploadMetadataInput:
      type: object
      properties:
        conversationId:
          type: string
    Upload:
      type: object
      required: [id, userId, filename, mimeType, sizeBytes, status, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        conversationId:
          type: ["string", "null"]
        filename:
          type: string
        originalFilename:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
        status:
          type: string
          enum: [processing, ready, failed]
        createdAt:
          type: string
          format: date-time

    # Settings schemas
...